managed/managed-config-success.https.html

<!DOCTYPE html>
<title>Managed Configuration: Returned values</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/test-only-api.js"></script>
<script src="resources/managed-configuration-helper.js"></script>
<script>
'use strict'

managed_config_test(async (test, managedConfigTest) => {
  promise_rejects_dom(test, 'NotAllowedError', navigator.managed.getManagedConfiguration(['a']));
}, "App is not managed.");

managed_config_test(async (test, managedConfigTest) => {
  managedConfigTest.setManagedConfig({"a" : '2'});
  const result = await navigator.managed.getManagedConfiguration(['a']);
  assert_equals(Object.keys(result).length, 1);
  assert_equals(result.a, 2);
}, "Configuration is returned");

managed_config_test(async (test, managedConfigTest) => {
  managedConfigTest.setManagedConfig({"a" : '2', 'b' : '3', 'c' : '1'});
  const result = await navigator.managed.getManagedConfiguration(['b', 'c']);
  assert_equals(Object.keys(result).length, 2);
  assert_equals(result.b, 3);
  assert_equals(result.c, 1);
}, "Selected keys are returned.");

managed_config_test(async (test, managedConfigTest) => {
  managedConfigTest.setManagedConfig({"a" : '2', 'b' : '3', 'c' : '1'});
  const result = await navigator.managed.getManagedConfiguration(['b', 'e']);
  assert_equals(Object.keys(result).length, 1);
  assert_equals(result.b, 3);
}, "Only existing keys are returned.");


managed_config_test(async (test, managedConfigTest) => {
  await new Promise(async resolve => {
    navigator.managed.addEventListener('managedconfigurationchange', function () {
      resolve();
    });
    await managedConfigTest.waitTillObserverAdded();
    managedConfigTest.setManagedConfig({"a" : '2'});
  });
}, "A change in managed configuration is observed.");



</script>
